🔨 Refactor: qwen-vl 관련 로직 ai 패키지로 이동
